index .ejs
*Fix the Add button:
Replace the empty href attribute of the Add button with the correct route to the BookDetails page (views/books/details.ejs). You need to provide the appropriate route for the desired functionality. Here's an example of how you can set the href attribute:

<a href="/books/details" class="btn btn-primary"><i class="fa fa-plus"></i> Add a book</a>

*Add the Edit button:
Inside the table row where you want to add the Edit button, insert the following code in the respective <td> element:

<td class="text-center">
  <a href="/books/edit/<%= books[count]._id %>" class="btn btn-primary btn-sm"><i class="fa fa-pencil"></i> Edit</a>
</td>

This code creates a link with the book's _id passed as a parameter in the URL, allowing the BookDetails page to identify which book to edit.

*Add the Delete button:

Inside the table row where you want to add the Delete button, insert the following code in the respective <td> element:

<td class="text-center">
  <a href="/books/delete/<%= books[count]._id %>" class="btn btn-danger btn-sm"><i class="fa fa-trash"></i> Delete</a>
</td>